requirements.txt
alembic==1.17.1
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-Scss==0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pyScss==1.4.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
python-dotenv==1.0.0
